Server works now.
Added mentor_request and mentor_topic. Needed to change composite primary key on them in database to a singular primary key. Updated schema file accordingly. Updated README with instructions on setting up the project, including the database. Tested and seems to be working. Added setup env vars script with placeholders. Going to ignore it next commit. Need to check out what's going on in the commentsRouter.